Categories are implemented with the structure, not a metadata.
Updates bootstrap to the latest version.

Adds a category for Valum since I will have to post about the
development of the framework quite often.

Link to the category in post that have a one.

The project layout is designed to show a project description and posts
in the category.