Layout improvments
The index.html display full post content, excerpts are for the archives that
display literally all posts in the site.

Layout to present any category rather than a project. It's okay to present all
posts as they should not be too numerous.

Links to categories are automatically generated in the sidebar menu. Only
categories having posts are shown.

Paragraphs in categories, posts and pages are justified for a better reading
experience.